update tracker and add tooltip for table
