+++ cloud/blockstore/libs/storage/partition/part_actor_writeblob.cpp
@@ -393,7 +393,8 @@ void TPartitionActor::HandleWriteBlob(
 
     GroupOperationTimeTracker.OnStarted(
         requestInfo->CallContext->RequestId,
-        TStringBuilder() << "Write_" << groupId,
+        groupId,
+        TGroupOperationTimeTracker::EGroupOperationType::Write,
         GetCycleCount());
 
     State->EnqueueIORequest(
